+++ lib/flutter_monetization_kit.dart
@@ -1,7 +1,0 @@
-import 'flutter_monetization_kit_platform_interface.dart';
-
-class FlutterMonetizationKit {
-  Future<String?> getPlatformVersion() {
-    return FlutterMonetizationKitPlatform.instance.getPlatformVersion();
-  }
-}