+++ lib/src/google_ads/widgets/native_widget.dart
@@ -192,7 +192,7 @@ class _NativeWidgetState extends State<NativeWidget> {
 
     String colorToHex(Color? color) {
       if (color == null) return '';
-      return '#${color.value.toRadixString(16).padLeft(8, '0')}';
+      return '#${color.toARGB32().toRadixString(16).padLeft(8, '0')}';
     }
 
     final Map<String, dynamic> creationParams = <String, dynamic>{