+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Fixed deprecation warnings for `withOpacity` by using `withValues`.
+* Fixed deprecation warnings for `Color.value` by using `toARGB32()`.
+* Cleaned up boilerplate test code in Dart and Kotlin.
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release of the Flutter Monetization Kit.